Refactor ArrayContainer and FITSLoader
The loader is now passed a reference to the array container which it uses to
lookup the basepath at load time.

In addition this commit also removes a lot of caching and other clutter, this is
now left to Dask / future work, as it would be worse to over cache rather than
under cache.